Split hard.py into the four things it was doing
669 lines in one module covering the on-disk layout, label preparation, the gate, and the four-step
training run. Now:

  hard/layout.py  the file names, and has_hard_head — pathlib and nothing else
  hard/labels.py  prepare_hard_labels_wide: your measurements -> hard.h5
  hard/gate.py    T and the blend ceiling — a_max, and how near your chemistry a query is
  hard/train.py   the four steps of learn-hard
  hard/__init__.py  the module docstring, and the same public surface as before

Every body moved verbatim, sliced by line range rather than retyped. `olinda.hard`'s public surface is
byte-identical to the old module's, so cli.py, export.py and the tests are untouched — export.py now
names `olinda.hard.layout` explicitly, which is the only import that changed.

One thing worth being clear about, since layout.py's docstring could imply otherwise: it is
dependency-free as a *module*, but reaching it as `olinda.hard.layout` still runs the package
`__init__`, which imports the rest. So export.py does not pay less than it did — it imports numpy
itself three lines further down, and is only ever imported from inside a CLI command body. The value
here is that the on-disk contract is stated in one place that has no reason to change.
